+++ CHANGELOG.md
@@ -14,6 +14,10 @@ The format is based on [Keep a Changelog].
   Minecraft 26.2).
 - Name Tags are not nameable by default through Anvil.
   This behavior is configurable.
+- Pigeons will now fly to their delivery target if within pathfinding range.
+  If their target falls outside of pathfinding range, Pigeons will fly towards their target's
+  general direction before teleporting to within pathfinding range of the target.
+  At that point, they will continue to fly to their target.
 
 ### Deprecated